Renamed `protodispatcher.rs` to `proto_dispatcher.rs` and adjusted related imports and module references. Added a `Sender` type alias for improved readability and simplified usage.
